Move `SessionSwitcher` & `CreateSession` to main App
* Update **browser** `App`: remove **sessionActor** initialization (src/browser/App.tsx);
* Update **main** `App`: add **sessionActor** initialization (src/main/App.tsx);
* Update **browser** `Layout`: remove **SessionSwitcher** & **sessionActor** state (src/browser/components/layout/Layout.tsx);
* Update **main** `Layout`: add **sessionActor** state (src/main/components/layout/Layout.tsx);
* Update **main** `Header`: add **CreateSession** dialog; add **SessionSwitcher** (src/main/components/common/Header/Header.tsx);
* Move `SessionSwitcher` from **browser** to ** main** _App_ (src/main/components/layout/Layout.tsx, src/main/components/templates/index.ts, src/main/components/templates/SessionSwitcher/index.ts, src/main/components/templates/SessionSwitcher/SessionSwitcher.tsx);
* Move `CreateSession` from **browser** to ** main** _App_ : (src/main/components/dialogs/CreateSession/CreateSession.tsx, src/main/components/dialogs/CreateSession/index.ts, src/main/components/dialogs/index.ts);
* Update node dependencies (yarn.lock, package-lock.json);
* Clear **vite** build on `npm start` (package.json);
* Remove **optimizeDeps** from `vite.main_window.config.ts` & `vite.browser_window.config.ts` (config/vite.main_window.config.ts, config/vite.browser_window.config.ts).
