+++ config/vite.main_window.config.ts
@@ -23,8 +23,5 @@ export default defineConfig((env) => {
       preserveSymlinks: true,
     },
     clearScreen: false,
-    optimizeDeps: {
-      exclude: ["blip-ds/loader"],
-    },
   } as UserConfig;
 });
