--- src/browser/components/dialogs/index.ts
@@ -1,1 +1,0 @@
-export { default as CreateSession } from "./CreateSession";
