+++ src/browser/components/templates/index.ts
@@ -1,3 +1,2 @@
 export { default as SearchBar } from "./SearchBar";
-export { default as SessionSwitcher } from "./SessionSwitcher";
 export { default as TabsBar } from "./TabsBar";
