+++ src/main/components/dialogs/CreateSession/index.ts
