+++ src/main/components/dialogs/index.ts
@@ -1,4 +1,5 @@
 export { default as CreateProxy } from "./CreateProxy";
+export { default as CreateSession } from "./CreateSession";
 export { default as DeleteProxy } from "./DeleteProxy";
 export { default as EditProxy } from "./EditProxy";
 export { default as Settings } from "./Settings";
