+++ src/main/components/templates/index.ts
@@ -3,3 +3,4 @@ export { default as ProxiesList } from "./ProxiesList";
 export { default as ProxyCard } from "./ProxyCard";
 export { default as ProxyForm } from "./ProxyForm";
 export { default as ProxyLog } from "./ProxyLog";
+export { default as SessionSwitcher } from "./SessionSwitcher";
